Moved `Renderer` into its own file
